+++ duckscript_sdk/src/sdk/std/env/mod.rs
@@ -1,10 +1,15 @@
 mod get;
 mod get_home_dir;
 mod get_user_name;
+mod is_windows;
 mod os_family;
+mod os_name;
+mod os_release;
+mod os_version;
 mod print_current_directory;
 mod set;
 mod set_current_directory;
+mod uname;
 mod unset;
 
 use crate::utils::pckg;
@@ -18,11 +23,16 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
 
     commands.set(get::create(&package))?;
     commands.set(get_home_dir::create(&package))?;
-    commands.set(get_user_name::create(PACKAGE))?;
-    commands.set(os_family::create(PACKAGE))?;
+    commands.set(get_user_name::create(&package))?;
+    commands.set(is_windows::create(&package)?)?;
+    commands.set(os_family::create(&package))?;
+    commands.set(os_name::create(&package))?;
+    commands.set(os_release::create(&package))?;
+    commands.set(os_version::create(&package))?;
     commands.set(print_current_directory::create(&package))?;
     commands.set(set::create(&package))?;
     commands.set(set_current_directory::create(&package))?;
+    commands.set(uname::create(&package)?)?;
     commands.set(unset::create(&package))?;
 
     Ok(())