+++ duckscript_sdk/src/sdk/std/env/uname/mod_test.rs
@@ -1,0 +1,7 @@
+use super::*;
+use crate::test;
+
+#[test]
+fn common_functions() {
+    test::test_common_command_functions(create("").unwrap());
+}