+++ duckscript_sdk/src/sdk/std/lib/alias/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod set;
+mod set;
 mod unset;
 
 use crate::utils::pckg;