+++ duckscript_sdk/src/sdk/std/lib/command/remove/mod.rs
@@ -1,6 +1,4 @@
-use crate::sdk::std::alias::ALIAS_STATE_KEY;
 use crate::utils::pckg;
-use crate::utils::state::get_sub_state;
 use duckscript::types::command::{Command, CommandResult, Commands};
 use duckscript::types::instruction::Instruction;
 use duckscript::types::runtime::StateValue;
@@ -17,11 +15,11 @@ pub(crate) struct CommandImpl {
 
 impl Command for CommandImpl {
     fn name(&self) -> String {
-        pckg::concat(&self.package, "Unset")
+        pckg::concat(&self.package, "Remove")
     }
 
     fn aliases(&self) -> Vec<String> {
-        vec!["unalias".to_string()]
+        vec!["remove_command".to_string()]
     }
 
     fn help(&self) -> String {
@@ -39,7 +37,7 @@ impl Command for CommandImpl {
     fn run_with_context(
         &self,
         arguments: Vec<String>,
-        state: &mut HashMap<String, StateValue>,
+        _state: &mut HashMap<String, StateValue>,
         _variables: &mut HashMap<String, String>,
         _output_variable: Option<String>,
         _instructions: &Vec<Instruction>,
@@ -47,25 +45,9 @@ impl Command for CommandImpl {
         _line: usize,
     ) -> CommandResult {
         if arguments.len() != 1 {
-            CommandResult::Error("Invalid alias name provided.".to_string())
+            CommandResult::Error("Invalid command name provided.".to_string())
         } else {
-            let sub_state = get_sub_state(ALIAS_STATE_KEY.to_string(), state);
-
-            let key = &arguments[0];
-            let removed = if sub_state.contains_key(key) {
-                if commands.remove(key) {
-                    sub_state.remove(key);
-                    true
-                } else {
-                    false
-                }
-            } else if commands.aliases.contains_key(key) {
-                commands.aliases.remove(key);
-                true
-            } else {
-                false
-            };
-
+            let removed = commands.remove(&arguments[0]);
             CommandResult::Continue(Some(removed.to_string()))
         }
     }